feat(nvim): rename module
