+++ config/nvim/lua/je12emy/plugins/languages.lua
@@ -4,7 +4,7 @@ return {
         lazy = true,
         ft = "rs",
         config = function()
-            local tools = require 'je12emy/lsp/lsp-tools'
+            local tools = require 'je12emy.utils.lsp-tools'
             local capabilities = tools.get_capabilities()
             require("rust_tools").setup({
                 dap = {
@@ -21,25 +21,24 @@ return {
             })
         end
     }, {
-        "jose-elias-alvarez/typescript.nvim",
-        ft = {"ts", "js", "tsx", "jsx"},
-        lazy = true,
-        config = function()
-            local tools = require 'je12emy.lsp.lsp-tools'
-            require("typescript").setup({
-                disable_commands = false, -- prevent the plugin from creating Vim commands
-                debug = false, -- enable debug logging for commands
-                go_to_source_definition = {
-                    fallback = true -- fall back to standard LSP definition on failure
-                },
-                server = { -- pass options to lspconfig's setup method
-                    on_attach = tools.on_attach
-                }
-            })
-
-        end
-    }, {
-        "wuelnerdotexe/vim-astro",
-        config = function() vim.g.astro_typescript = 'enable' end
-    }
+    "jose-elias-alvarez/typescript.nvim",
+    ft = { "ts", "js", "tsx", "jsx" },
+    lazy = true,
+    config = function()
+        local tools = require 'je12emy.utils.lsp-tools'
+        require("typescript").setup({
+            disable_commands = false, -- prevent the plugin from creating Vim commands
+            debug = false,            -- enable debug logging for commands
+            go_to_source_definition = {
+                fallback = true       -- fall back to standard LSP definition on failure
+            },
+            server = {                -- pass options to lspconfig's setup method
+                on_attach = tools.on_attach
+            }
+        })
+    end
+}, {
+    "wuelnerdotexe/vim-astro",
+    config = function() vim.g.astro_typescript = 'enable' end
+}
 }
